add styles
