+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { useState, Fragment} from 'react';
 import './App.scss';
-import './components/contentSection/SectionStyles.scss';
+import './components/sections/SectionStyles.scss';
 import './components/modals/PageModals.scss';
 import Header from './components/header/Header';
 import Menu from './components/menu/Menu';
 import ContentContainer from './components/contentContainer/ContentContainer';
 import Footer from './components/footer/Footer';
-import Radio from './components/contexts/RadioContext';
+import Radio from './contexts/RadioContext';
 
 const App = () => {
 	const [radio, setRadio] = useState("characters")
