+++ src/components/footer/Footer.scss
@@ -2,7 +2,7 @@
 
 footer {
     align-items: center;
-    background-color: #B3C100;
+    background-color: #4Cb5F5;
     display: flex;
     height: 50px;
     justify-content: center;
